requirements.txt
psycopg2
asgiref
Django
pytz
sqlparse
Pillow
celery
redis
flower
djangorestframework
django-filter
requests
drf-spectacular